@@ -4,43 +4,24 @@
   "private": true,
   "homepage": "http://esaleshome.com/ads",
   "dependencies": {
-    "babel-preset-env": "^1.7.0",
-    "babel-preset-react": "^6.24.1",
     "custom-react-scripts": "^0.2.2",
-    "enzyme": "^3.7.0",
-    "enzyme-adapter-react-16": "^1.7.0",
-    "jest-svg-transformer": "^1.0.0",
-    "react": "^16.4.1",
-    "react-autocomplete": "^1.8.1",
-    "react-autosuggest": "^9.4.3",
+    "react": "^16.12.0",
     "react-bootstrap": "^0.32.4",
     "react-cookie": "^2.2.0",
-    "react-device-detect": "^1.6.1",
-    "react-dom": "^16.4.1",
+    "react-dom": "^16.12.0",
     "react-redux": "^5.1.0",
-    "react-responsive": "^6.0.1",
-    "react-router-dom": "^4.3.1",
-    "react-sanitized-html": "^2.0.0",
+    "react-router-dom": "^5.1.2",
     "react-scripts": "1.1.4",
     "redux": "^4.0.1",
-    "redux-mock-store": "^1.5.3",
     "redux-saga": "^0.16.2",
-    "redux-saga-effects": "^1.1.0",
-    "sanitize-html": "^1.18.2"
+    "redux-saga-effects": "^1.1.0"
   },
   "scripts": {
     "start": "react-scripts start",
     "build": "react-scripts build",
-    "test": "jest",
     "eject": "react-scripts eject"
   },
   "eslintConfig": {
     "extends": "react-app"
-  },
-  "devDependencies": {
-    "babel-core": "^6.26.3",
-    "babel-jest": "^23.6.0",
-    "jest": "^23.6.0",
-    "regenerator-runtime": "^0.12.1"
   }
 }
