@@ -1,8 +1,8 @@
 import React, { Component } from 'react';
 import { connect } from 'react-redux';
-import { Button	} from 'react-bootstrap';
+import { Button } from 'react-bootstrap';
 import { getLabel } from '../common/label';
-import { AdFuncs} from '../common/ad-funcs';
+import { AdFuncs } from '../common/ad-funcs';
 import { Helper } from '../common/js-funcs';
 import { Engine } from '../common/engine';
 import UserTabs from '../user/user-tabs';
@@ -19,14 +19,14 @@ import {
 } from '../action';
 
 class EditAdvertise extends Component {
-	componentWillMount() {
+	componentDidMount() {
 		const advertiseId = Engine.param('advertiseId') || null;
 
-        AdFuncs.initOtherItems();
+		AdFuncs.initOtherItems();
 
-		Engine.dispatch(fetchUserData([{dataKey: 'profile'}]));
-		
-        if(this.props.USERSIGNEDIN && advertiseId) {
+		Engine.dispatch(fetchUserData([{ dataKey: 'profile' }]));
+
+		if (this.props.USERSIGNEDIN && advertiseId) {
 			Engine.dispatch(fetchUserAdvertise(advertiseId));
 		}
 	}
@@ -34,207 +34,207 @@ class EditAdvertise extends Component {
 	componentWillUnmount() {
 		Engine.dispatch(clearStateModelData());
 	}
-	
+
 	render() {
-		const {POSTERRORS, MODELS, profile} = this.props,
+		const { POSTERRORS, MODELS, profile } = this.props,
 			ad = MODELS['userAdvertiseSet'];
 
-		if(!ad || !ad.id) {
+		if (!ad || !ad.id) {
 			return null;
-		}else {
+		} else {
 			AdFuncs.initCity(MODELS['cityId']);
 			AdFuncs.initItem(MODELS['itemId']);
 			AdFuncs.initImages(MODELS['images']);
 			AdFuncs.initContactMethods(MODELS['contactMethods'], MODELS['_contactMethods']);
 		}
-		
-        return (
+
+		return (
 			<React.Fragment>
-				{Engine.component(UserTabs, {profile})}
+				{Engine.component(UserTabs, { profile })}
 				{
-					this.props.USERSIGNEDIN ? 
-					(<form name="work_on_advertise" action="edit-ad" encType="multipart/form-data"> 
-						<div className="ad_inline_block"><div className="ad_box_title">{getLabel('p_edit_title')}<div className="ang_close icon_list" onClick={Helper.goBack}>&times;</div></div></div><div className="ad_inline_block_top"></div>
-						<div className="ad_inline_block">
-							<div className="ad_box_field"><div className="ad_box_label">{getLabel('p_cat_title')}</div></div>
-							<div className="ad_box_field">
-								{Helper.errorLabel(POSTERRORS.category)}
-								{Engine.component(CategoryList)}
-							</div>
-						</div><div className="ad_inline_block_top">
-							<div className="ad_box_field">
-								<div className="post_hints">
-									{getLabel('p_cat_hint')}
+					this.props.USERSIGNEDIN ?
+						(<form name="work_on_advertise" action="edit-ad" encType="multipart/form-data">
+							<div className="ad_inline_block"><div className="ad_box_title">{getLabel('p_edit_title')}<div className="ang_close icon_list" onClick={Helper.goBack}>&times;</div></div></div><div className="ad_inline_block_top"></div>
+							<div className="ad_inline_block">
+								<div className="ad_box_field"><div className="ad_box_label">{getLabel('p_cat_title')}</div></div>
+								<div className="ad_box_field">
+									{Helper.errorLabel(POSTERRORS.category)}
+									{Engine.component(CategoryList)}
 								</div>
-							</div>
-						</div>
-						<div className="ad_inline_block">
-							<div className="ad_box_field"><div className="ad_box_label">{getLabel('p_sub_cat_title')}</div></div>
-							<div className="ad_box_field">
-								{Helper.errorLabel(POSTERRORS.item)}
-								{Engine.component(ItemList)}
-							</div>
-						</div><div className="ad_inline_block_top">
-							<div className="ad_box_field">
-								<div className="post_hints">
-									{getLabel('p_sub_cat_hint')}
-								</div>
-							</div>
-						</div>
-						<div className="ad_inline_block">
-							<div className="ad_box_field"><div className="ad_box_label">{getLabel('p_name_title')}</div></div>
-							<div className="ad_box_field">
-								{Helper.errorLabel(POSTERRORS.ad_name)}
-								<input type="text" name="ad_name" defaultValue={ad.name} maxLength="100" />
-							</div>
-						</div><div className="ad_inline_block_top">
-							<div className="ad_box_field">
-								<div className="post_hints">
-									{getLabel('p_ad_name_hint')}
-								</div>
-							</div>
-						</div>
-						<div className="ad_inline_block">
-							<div className="ad_box_field"><div className="ad_box_label">{getLabel('p_free_title')}</div></div>
-							<div className="ad_box_field">
-								{Helper.errorLabel(POSTERRORS.is_free)}
-								<nav className="btn_list">
-									<div className={`tab ${AdFuncs.itemIsFree() ? 'active' : ''}`} onClick={AdFuncs.adIsFree}>{getLabel('p_yes_title')}</div>
-									<div className={`tab ${AdFuncs.itemIsNotFree() ? 'active' : ''}`} onClick={AdFuncs.adNotFree}>{getLabel('p_no_title')}</div>
-								</nav>
-								<input type="hidden" name="is_free" defaultValue={MODELS['is_free']} />
-							</div>
-						</div><div className="ad_inline_block_top">
-							<div className="ad_box_field">
-								<div className="post_hints">
-								</div>
-							</div>
-						</div>
-						{
-							AdFuncs.itemIsNotFree() ? (
-								<div id="price_row">
-									<div className="ad_inline_block">
-										<div className="ad_box_field"><div className="ad_box_label">{getLabel('p_currency_title')}</div></div>
-										<div className="ad_box_field">
-											{Helper.errorLabel(POSTERRORS.currency)}
-											<nav className="btn_list">
-												<div className={`tab ${MODELS['currency'] === 'c' ? 'active' : ''}`} onClick={() => {AdFuncs.setCurrency('c')}}>{getLabel('p_can_dollar')}</div>
-												<div className={`tab ${MODELS['currency'] === 'u' ? 'active' : ''}`} onClick={() => {AdFuncs.setCurrency('u')}}>{getLabel('p_us_dollar')}</div>
-												<div className={`tab ${MODELS['currency'] === 'r' ? 'active' : ''}`} onClick={() => {AdFuncs.setCurrency('r')}}>{getLabel('p_rmb_yuan')}</div>
-											</nav>
-											<input type="hidden" name="currency" defaultValue={MODELS['currency']} />
-										</div>
-									</div><div className="ad_inline_block_top">
-									<div className="ad_box_field">
-										<div className="currency_hints">
-										</div>
+							</div><div className="ad_inline_block_top">
+								<div className="ad_box_field">
+									<div className="post_hints">
+										{getLabel('p_cat_hint')}
 									</div>
 								</div>
-									<div className="ad_inline_block">
-										<div className="ad_box_field"><div className="ad_box_label">{getLabel('p_price_title')}</div></div>
-										<div className="ad_box_field">
-											{Helper.errorLabel(POSTERRORS.price)}
-											<input type="text" name="price" defaultValue={MODELS['price']} maxLength="15" />
-										</div>
-									</div><div className="ad_inline_block_top">
-									<div className="ad_box_field">
-										<div className="post_hints">
-											{getLabel('p_price_hint')}
-										</div>
+							</div>
+							<div className="ad_inline_block">
+								<div className="ad_box_field"><div className="ad_box_label">{getLabel('p_sub_cat_title')}</div></div>
+								<div className="ad_box_field">
+									{Helper.errorLabel(POSTERRORS.item)}
+									{Engine.component(ItemList)}
+								</div>
+							</div><div className="ad_inline_block_top">
+								<div className="ad_box_field">
+									<div className="post_hints">
+										{getLabel('p_sub_cat_hint')}
 									</div>
 								</div>
-								</div>
-							) : null
-						}
-						<div className="ad_inline_block">
-							<div className="ad_box_field"><div className="ad_box_label">{getLabel('p_detail_title')}</div></div>
-							<div className="ad_box_field">
-								<textarea name="description" rows="4" cols="100" placeholder={getLabel('p_detail_placeholder')} defaultValue={ad.description} />
 							</div>
-						</div><div className="ad_inline_block_top">
-							<div className="ad_box_field">
-								<div className="post_hints">
-									{getLabel('p_detail_hint')}
+							<div className="ad_inline_block">
+								<div className="ad_box_field"><div className="ad_box_label">{getLabel('p_name_title')}</div></div>
+								<div className="ad_box_field">
+									{Helper.errorLabel(POSTERRORS.ad_name)}
+									<input type="text" name="ad_name" defaultValue={ad.name} maxLength="100" />
+								</div>
+							</div><div className="ad_inline_block_top">
+								<div className="ad_box_field">
+									<div className="post_hints">
+										{getLabel('p_ad_name_hint')}
+									</div>
 								</div>
 							</div>
-						</div>
-						<div className="ad_inline_block">
-							<div className="ad_box_field">
+							<div className="ad_inline_block">
+								<div className="ad_box_field"><div className="ad_box_label">{getLabel('p_free_title')}</div></div>
+								<div className="ad_box_field">
+									{Helper.errorLabel(POSTERRORS.is_free)}
+									<nav className="btn_list">
+										<div className={`tab ${AdFuncs.itemIsFree() ? 'active' : ''}`} onClick={AdFuncs.adIsFree}>{getLabel('p_yes_title')}</div>
+										<div className={`tab ${AdFuncs.itemIsNotFree() ? 'active' : ''}`} onClick={AdFuncs.adNotFree}>{getLabel('p_no_title')}</div>
+									</nav>
+									<input type="hidden" name="is_free" defaultValue={MODELS['is_free']} />
+								</div>
+							</div><div className="ad_inline_block_top">
+								<div className="ad_box_field">
+									<div className="post_hints">
+									</div>
+								</div>
+							</div>
 							{
-								AdFuncs.adImageFull() ?
-								(<div className="tab inactive">{getLabel('r_add_photos')}</div>) :
-								(<div className="tab" onClick={AdFuncs.addAdImage}>{getLabel('r_add_photos')}</div>)
+								AdFuncs.itemIsNotFree() ? (
+									<div id="price_row">
+										<div className="ad_inline_block">
+											<div className="ad_box_field"><div className="ad_box_label">{getLabel('p_currency_title')}</div></div>
+											<div className="ad_box_field">
+												{Helper.errorLabel(POSTERRORS.currency)}
+												<nav className="btn_list">
+													<div className={`tab ${MODELS['currency'] === 'c' ? 'active' : ''}`} onClick={() => { AdFuncs.setCurrency('c') }}>{getLabel('p_can_dollar')}</div>
+													<div className={`tab ${MODELS['currency'] === 'u' ? 'active' : ''}`} onClick={() => { AdFuncs.setCurrency('u') }}>{getLabel('p_us_dollar')}</div>
+													<div className={`tab ${MODELS['currency'] === 'r' ? 'active' : ''}`} onClick={() => { AdFuncs.setCurrency('r') }}>{getLabel('p_rmb_yuan')}</div>
+												</nav>
+												<input type="hidden" name="currency" defaultValue={MODELS['currency']} />
+											</div>
+										</div><div className="ad_inline_block_top">
+											<div className="ad_box_field">
+												<div className="currency_hints">
+												</div>
+											</div>
+										</div>
+										<div className="ad_inline_block">
+											<div className="ad_box_field"><div className="ad_box_label">{getLabel('p_price_title')}</div></div>
+											<div className="ad_box_field">
+												{Helper.errorLabel(POSTERRORS.price)}
+												<input type="text" name="price" defaultValue={MODELS['price']} maxLength="15" />
+											</div>
+										</div><div className="ad_inline_block_top">
+											<div className="ad_box_field">
+												<div className="post_hints">
+													{getLabel('p_price_hint')}
+												</div>
+											</div>
+										</div>
+									</div>
+								) : null
 							}
-							</div>
-							{Engine.component(ImageList)}
-						</div><div className="ad_inline_block_top">
-							<div className="ad_box_field">
-								<div className="post_hints">
-									{getLabel('p_add_photo_hint')} 
+							<div className="ad_inline_block">
+								<div className="ad_box_field"><div className="ad_box_label">{getLabel('p_detail_title')}</div></div>
+								<div className="ad_box_field">
+									<textarea name="description" rows="4" cols="100" placeholder={getLabel('p_detail_placeholder')} defaultValue={ad.description} />
+								</div>
+							</div><div className="ad_inline_block_top">
+								<div className="ad_box_field">
+									<div className="post_hints">
+										{getLabel('p_detail_hint')}
+									</div>
 								</div>
 							</div>
-						</div>
-						<div className="ad_inline_block">
-							<div className="ad_box_field"><div className="ad_box_label">{getLabel('p_contact_title')}</div></div>
-							<div className="ad_box_field">
-								{Helper.errorLabel(POSTERRORS.contact_method)}
-								<nav className="btn_list">
-									<div className={`tab ${AdFuncs.isContactMethod('contact_phone') ? 'active' : ''}`} onClick={() => {AdFuncs.setContactMethod('contact_phone')}}>{getLabel('p_phone')}</div>
-									<div className={`tab ${AdFuncs.isContactMethod('contact_email') ? 'active' : ''}`} onClick={() => {AdFuncs.setContactMethod('contact_email')}}>{getLabel('p_email')}</div>
-								</nav>
-								<input type="hidden" name="contact_method" defaultValue={AdFuncs.contactMethod()} />
-							</div>
-						</div><div className="ad_inline_block_top">
-							<div className="ad_box_field">
-								<div className="post_hints">
-									{getLabel('p_contact_you_title')}
+							<div className="ad_inline_block">
+								<div className="ad_box_field">
+									{
+										AdFuncs.adImageFull() ?
+											(<div className="tab inactive">{getLabel('r_add_photos')}</div>) :
+											(<div className="tab" onClick={AdFuncs.addAdImage}>{getLabel('r_add_photos')}</div>)
+									}
+								</div>
+								{Engine.component(ImageList)}
+							</div><div className="ad_inline_block_top">
+								<div className="ad_box_field">
+									<div className="post_hints">
+										{getLabel('p_add_photo_hint')}
+									</div>
 								</div>
 							</div>
-						</div>
-						{Engine.component(ContactPhone, {MODELS, POSTERRORS})}
-						<div className="ad_inline_block">
-							<div className="ad_box_field"><div className="ad_box_label">{getLabel('p_province_title')}</div></div>
-							<div className="ad_box_field">
-								{Helper.errorLabel(POSTERRORS.province)}
-								{Engine.component(ProvinceList)}
-							</div>
-						</div><div className="ad_inline_block_top">
-							<div className="ad_box_field">
-								<div className="post_hints">
-									{getLabel('p_province_hint')}
+							<div className="ad_inline_block">
+								<div className="ad_box_field"><div className="ad_box_label">{getLabel('p_contact_title')}</div></div>
+								<div className="ad_box_field">
+									{Helper.errorLabel(POSTERRORS.contact_method)}
+									<nav className="btn_list">
+										<div className={`tab ${AdFuncs.isContactMethod('contact_phone') ? 'active' : ''}`} onClick={() => { AdFuncs.setContactMethod('contact_phone') }}>{getLabel('p_phone')}</div>
+										<div className={`tab ${AdFuncs.isContactMethod('contact_email') ? 'active' : ''}`} onClick={() => { AdFuncs.setContactMethod('contact_email') }}>{getLabel('p_email')}</div>
+									</nav>
+									<input type="hidden" name="contact_method" defaultValue={AdFuncs.contactMethod()} />
+								</div>
+							</div><div className="ad_inline_block_top">
+								<div className="ad_box_field">
+									<div className="post_hints">
+										{getLabel('p_contact_you_title')}
+									</div>
 								</div>
 							</div>
-						</div>
-						<div className="ad_inline_block">
-							<div className="ad_box_field"><div className="ad_box_label">{getLabel('p_city_title')}</div></div>
-							<div className="ad_box_field">
-								{Helper.errorLabel(POSTERRORS.city)}
-								{Engine.component(CityList)}
-							</div>
-						</div><div className="ad_inline_block_top">
-							<div className="ad_box_field">
-								<div className="post_hints">
-									{getLabel('p_city_hint')}
+							{Engine.component(ContactPhone, { MODELS, POSTERRORS })}
+							<div className="ad_inline_block">
+								<div className="ad_box_field"><div className="ad_box_label">{getLabel('p_province_title')}</div></div>
+								<div className="ad_box_field">
+									{Helper.errorLabel(POSTERRORS.province)}
+									{Engine.component(ProvinceList)}
+								</div>
+							</div><div className="ad_inline_block_top">
+								<div className="ad_box_field">
+									<div className="post_hints">
+										{getLabel('p_province_hint')}
+									</div>
 								</div>
 							</div>
-						</div>
-						<div className="ad_inline_block">
-							<div className="ad_box_field"><div className="ad_box_label">{getLabel('p_add_title')}</div></div>
-							<div className="ad_box_field">
-								<input type="text" name="address" defaultValue={ad.address} maxLength="100" placeholder={getLabel('p_add_placeholder')} />
-							</div>
-						</div><div className="ad_inline_block_top">
-							<div className="ad_box_field">
-								<div className="post_hints">
-									{getLabel('p_add_hint')}
+							<div className="ad_inline_block">
+								<div className="ad_box_field"><div className="ad_box_label">{getLabel('p_city_title')}</div></div>
+								<div className="ad_box_field">
+									{Helper.errorLabel(POSTERRORS.city)}
+									{Engine.component(CityList)}
+								</div>
+							</div><div className="ad_inline_block_top">
+								<div className="ad_box_field">
+									<div className="post_hints">
+										{getLabel('p_city_hint')}
+									</div>
 								</div>
 							</div>
-						</div>
-						<div className="ad_box_field"><div className="ad_box_label"></div>
-							<Button bsStyle={Object.keys(POSTERRORS).length ? 'danger' : 'success'} bsSize='small' onClick={() => Helper.submitForm('work_on_advertise', true)}>{getLabel('p_upt_post')}</Button>
-						</div>
-						<input type="hidden" name="advertise_id" defaultValue={ad.id} />
-					</form>) : null
+							<div className="ad_inline_block">
+								<div className="ad_box_field"><div className="ad_box_label">{getLabel('p_add_title')}</div></div>
+								<div className="ad_box_field">
+									<input type="text" name="address" defaultValue={ad.address} maxLength="100" placeholder={getLabel('p_add_placeholder')} />
+								</div>
+							</div><div className="ad_inline_block_top">
+								<div className="ad_box_field">
+									<div className="post_hints">
+										{getLabel('p_add_hint')}
+									</div>
+								</div>
+							</div>
+							<div className="ad_box_field"><div className="ad_box_label"></div>
+								<Button bsStyle={Object.keys(POSTERRORS).length ? 'danger' : 'success'} bsSize='small' onClick={() => Helper.submitForm('work_on_advertise', true)}>{getLabel('p_upt_post')}</Button>
+							</div>
+							<input type="hidden" name="advertise_id" defaultValue={ad.id} />
+						</form>) : null
 				}
 			</React.Fragment>
 		);
