@@ -2,7 +2,7 @@ import React, { Component } from 'react';
 import { connect } from 'react-redux';
 import { Button, ButtonToolbar } from 'react-bootstrap';
 import { getLabel } from '../common/label';
-import { AdFuncs} from '../common/ad-funcs';
+import { AdFuncs } from '../common/ad-funcs';
 import { AdInfo } from '../common/ad-info';
 import { Helper } from '../common/js-funcs';
 import { Engine } from '../common/engine';
@@ -15,12 +15,12 @@ import ItemList from './item-list';
 import { fetchVisitorAdvertise, clearStateModelData } from '../action';
 
 class VisitorAdvertise extends Component {
-	componentWillMount() {
+	componentDidMount() {
 		let behavior = Engine.param('behavior') || null;
 
-        AdFuncs.initOtherItems();
+		AdFuncs.initOtherItems();
 
-        if(behavior === 'edit') {
+		if (behavior === 'edit') {
 			Engine.dispatch(fetchVisitorAdvertise());
 		}
 	}
@@ -28,28 +28,28 @@ class VisitorAdvertise extends Component {
 	componentWillUnmount() {
 		Engine.dispatch(clearStateModelData());
 	}
-	
-	render() {
-		const {POSTERRORS, MODELS} = this.props,
-            ad = MODELS['userAdvertiseSet'];
 
-		if(!ad || !ad.id) {
+	render() {
+		const { POSTERRORS, MODELS } = this.props,
+			ad = MODELS['userAdvertiseSet'];
+
+		if (!ad || !ad.id) {
 			return null;
-		}else {
+		} else {
 			AdFuncs.initCity(MODELS['cityId']);
 			AdFuncs.initItem(MODELS['itemId']);
 			AdFuncs.initImages(MODELS['images']);
 			AdFuncs.initContactMethods(MODELS['contactMethods'], MODELS['_contactMethods']);
 		}
-		
-        return (
-			<form name="work_on_advertise" action="visitor-ad" encType="multipart/form-data"> 
+
+		return (
+			<form name="work_on_advertise" action="visitor-ad" encType="multipart/form-data">
 				<div className="ad_inline_block"><div className="ad_box_title">{getLabel('p_edit_title')}<div className="ang_close icon_list" onClick={Helper.goBack}>&times;</div></div></div><div className="ad_inline_block_top">
 					<div className="ad_box_field">
 						<div className="post_hints">
 							<div>{getLabel('p_phone')} / {getLabel('p_email')}: {ad.contact_phone} / {ad.contact_email}</div>
 						</div>
-					</div>					
+					</div>
 				</div>
 				<div className="ad_inline_block">
 					<div className="ad_box_field"><div className="ad_box_label">{getLabel('p_cat_title')}</div></div>
@@ -114,18 +114,18 @@ class VisitorAdvertise extends Component {
 								<div className="ad_box_field">
 									{Helper.errorLabel(POSTERRORS.is_free)}
 									<nav className="btn_list">
-										<div className={`tab ${MODELS['currency'] === 'c' ? 'active' : ''}`} onClick={() => {AdFuncs.setCurrency('c')}}>{getLabel('p_can_dollar')}</div>
-										<div className={`tab ${MODELS['currency'] === 'u' ? 'active' : ''}`} onClick={() => {AdFuncs.setCurrency('u')}}>{getLabel('p_us_dollar')}</div>
-										<div className={`tab ${MODELS['currency'] === 'r' ? 'active' : ''}`} onClick={() => {AdFuncs.setCurrency('r')}}>{getLabel('p_rmb_yuan')}</div>
+										<div className={`tab ${MODELS['currency'] === 'c' ? 'active' : ''}`} onClick={() => { AdFuncs.setCurrency('c') }}>{getLabel('p_can_dollar')}</div>
+										<div className={`tab ${MODELS['currency'] === 'u' ? 'active' : ''}`} onClick={() => { AdFuncs.setCurrency('u') }}>{getLabel('p_us_dollar')}</div>
+										<div className={`tab ${MODELS['currency'] === 'r' ? 'active' : ''}`} onClick={() => { AdFuncs.setCurrency('r') }}>{getLabel('p_rmb_yuan')}</div>
 									</nav>
 									<input type="hidden" name="currency" defaultValue={MODELS['currency']} />
 								</div>
 							</div><div className="ad_inline_block_top">
-							<div className="ad_box_field">
-								<div className="currency_hints">
+								<div className="ad_box_field">
+									<div className="currency_hints">
+									</div>
 								</div>
 							</div>
-						</div>
 							<div className="ad_inline_block">
 								<div className="ad_box_field"><div className="ad_box_label">{getLabel('p_price_title')}</div></div>
 								<div className="ad_box_field">
@@ -133,12 +133,12 @@ class VisitorAdvertise extends Component {
 									<input type="text" name="price" defaultValue={MODELS['price']} maxLength="15" />
 								</div>
 							</div><div className="ad_inline_block_top">
-							<div className="ad_box_field">
-								<div className="post_hints">
-									{getLabel('p_price_hint')}
+								<div className="ad_box_field">
+									<div className="post_hints">
+										{getLabel('p_price_hint')}
+									</div>
 								</div>
 							</div>
-						</div>
 						</div>
 					) : null
 				}
@@ -156,17 +156,17 @@ class VisitorAdvertise extends Component {
 				</div>
 				<div className="ad_inline_block">
 					<div className="ad_box_field">
-					{
-						AdFuncs.adImageFull() ?
-						(<div className="tab inactive">{getLabel('r_add_photos')}</div>) :
-						(<div className="tab" onClick={AdFuncs.addAdImage}>{getLabel('r_add_photos')}</div>)
-					}
+						{
+							AdFuncs.adImageFull() ?
+								(<div className="tab inactive">{getLabel('r_add_photos')}</div>) :
+								(<div className="tab" onClick={AdFuncs.addAdImage}>{getLabel('r_add_photos')}</div>)
+						}
 					</div>
 					{Engine.component(ImageList)}
 				</div><div className="ad_inline_block_top">
 					<div className="ad_box_field">
 						<div className="post_hints">
-							{getLabel('p_add_photo_hint')} 
+							{getLabel('p_add_photo_hint')}
 						</div>
 					</div>
 				</div>
@@ -175,8 +175,8 @@ class VisitorAdvertise extends Component {
 					<div className="ad_box_field">
 						{Helper.errorLabel(POSTERRORS.contact_method)}
 						<nav className="btn_list">
-							<div className={`tab ${AdFuncs.isContactMethod('contact_phone') ? 'active' : ''}`} onClick={() => {AdFuncs.setContactMethod('contact_phone')}}>{getLabel('p_phone')}</div>
-							<div className={`tab ${AdFuncs.isContactMethod('contact_email') ? 'active' : ''}`} onClick={() => {AdFuncs.setContactMethod('contact_email')}}>{getLabel('p_email')}</div>
+							<div className={`tab ${AdFuncs.isContactMethod('contact_phone') ? 'active' : ''}`} onClick={() => { AdFuncs.setContactMethod('contact_phone') }}>{getLabel('p_phone')}</div>
+							<div className={`tab ${AdFuncs.isContactMethod('contact_email') ? 'active' : ''}`} onClick={() => { AdFuncs.setContactMethod('contact_email') }}>{getLabel('p_email')}</div>
 						</nav>
 						<input type="hidden" name="contact_method" defaultValue={AdFuncs.contactMethod()} />
 					</div>
@@ -187,7 +187,7 @@ class VisitorAdvertise extends Component {
 						</div>
 					</div>
 				</div>
-				{Engine.component(ContactPhone, {MODELS, POSTERRORS})}
+				{Engine.component(ContactPhone, { MODELS, POSTERRORS })}
 				<div className="ad_inline_block">
 					<div className="ad_box_field"><div className="ad_box_label">{getLabel('p_province_title')}</div></div>
 					<div className="ad_box_field">
@@ -227,10 +227,10 @@ class VisitorAdvertise extends Component {
 					</div>
 				</div>
 				<div className="ad_box_field"><div className="ad_box_label"></div>
-				<ButtonToolbar>
-					<Button bsStyle={Object.keys(POSTERRORS).length ? 'danger' : 'success'} onClick={() => Helper.submitForm('work_on_advertise', true)}>{getLabel('p_upt_post')}</Button>
-					<Button bsStyle='success' onClick={() => {AdInfo.visitorConfirmDeletion()}}>{getLabel('p_del_post')}</Button>
-				</ButtonToolbar>
+					<ButtonToolbar>
+						<Button bsStyle={Object.keys(POSTERRORS).length ? 'danger' : 'success'} onClick={() => Helper.submitForm('work_on_advertise', true)}>{getLabel('p_upt_post')}</Button>
+						<Button bsStyle='success' onClick={() => { AdInfo.visitorConfirmDeletion() }}>{getLabel('p_del_post')}</Button>
+					</ButtonToolbar>
 				</div>
 			</form>
 		);
