@@ -9,19 +9,19 @@ const UserAccount = class extends Component {
 	/*
 	 * @description user is already signed in and is now on the account page, he/she presses the F5 key
 	 */
-    componentWillMount() {
-        if(this.props.USERSIGNEDIN && !this.props.profile) {
-			Engine.dispatch(fetchUserData([{dataKey: 'profile'}]));
-    	}
-    }
+	componentDidMount() {
+		if (this.props.USERSIGNEDIN && !this.props.profile) {
+			Engine.dispatch(fetchUserData([{ dataKey: 'profile' }]));
+		}
+	}
 
 	render() {
-        const {profile} = this.props;
-			
+		const { profile } = this.props;
+
 		return (
 			<section>
-				{Engine.component(UserTabs, {profile, USERSIGNEDIN: this.props.USERSIGNEDIN})}
-				{Engine.component(UserProfile, {profile, USERSIGNEDIN: this.props.USERSIGNEDIN})}
+				{Engine.component(UserTabs, { profile, USERSIGNEDIN: this.props.USERSIGNEDIN })}
+				{Engine.component(UserProfile, { profile, USERSIGNEDIN: this.props.USERSIGNEDIN })}
 			</section>
 		);
 	};
