@@ -1,9 +1,6 @@
 import React, { Component } from 'react';
 import PropTypes from 'prop-types';
 import { connect } from 'react-redux';
-import MediaQuery from 'react-responsive';
-import logo from '../logo.svg';
-import node from '../node.png';
 import AdvertiseContact from './advertise-contact';
 import ContactUs from './contact-us';
 import EmailFriendBox from './email-friend-box';
@@ -22,15 +19,15 @@ import * as Action from '../action';
 class ActionBox extends Component {
 	goHome = () => {
 		AdInfo.listAdvertise();
-        Helper.location('');
+		Helper.location('');
 	};
 
 	componentDidMount() {
 		Engine.registerMenuItems();
 	}
-	
+
 	render() {
-		const {BOXES, ERRORS, LANGUAGE} = this.props;
+		const { BOXES, ERRORS, LANGUAGE } = this.props;
 
 		return (
 			<nav id="ad_title_bar">
@@ -39,30 +36,22 @@ class ActionBox extends Component {
 						{getLabel('c_menu')}
 					</div>
 					<div className="nav_btn" onClick={this.goHome}>
-                        {getLabel('c_home')}
+						{getLabel('c_home')}
 					</div>
-					<div className="nav_btn" onClick={() => {this.props.dispatch(switchLanguage())}}>
+					<div className="nav_btn" onClick={() => { this.props.dispatch(switchLanguage()) }}>
 						{getLabel('c_language')}
 					</div>
-					<MediaQuery minWidth={768}>
-						<div className="react-icon" title={getLabel('host_info')}>
-							<img className="rotate-on" alt="" src={logo} height="20" />
-						</div>
-						<div className="react-icon" title="NodeJs">
-							<img alt="" src={node} height="20" />
-						</div>
-					</MediaQuery>
 					<label htmlFor="keyWordSearch" className="visually-hidden">{getLabel('c_key_search')}</label>
 					<input name="keyWordSearch" id="keyWordSearch" className="key-word-search" placeholder={getLabel('c_key_search')} onChange={e => this.props.dispatch(Action.keyWordSearch(e.target.value))} />
 					<div className="ad_action_box_wrapper">
-						{Engine.component(AdvertiseContact, {BOXES, ERRORS, LANGUAGE})}
-						{Engine.component(ContactUs, {BOXES, ERRORS, LANGUAGE})}
-						{Engine.component(EmailFriendBox, {BOXES, ERRORS, LANGUAGE})}
-						{Engine.component(FindVisitorAdBox, {BOXES, ERRORS, LANGUAGE})}
-						{Engine.component(ForgetPasswordBox, {BOXES, ERRORS, LANGUAGE})}
-						{Engine.component(LoginBox, {BOXES, ERRORS, LANGUAGE})}
-						{Engine.component(Register, {BOXES, ERRORS, LANGUAGE})}
-						{Engine.component(SearchBox, {BOXES, ERRORS, LANGUAGE})}
+						{Engine.component(AdvertiseContact, { BOXES, ERRORS, LANGUAGE })}
+						{Engine.component(ContactUs, { BOXES, ERRORS, LANGUAGE })}
+						{Engine.component(EmailFriendBox, { BOXES, ERRORS, LANGUAGE })}
+						{Engine.component(FindVisitorAdBox, { BOXES, ERRORS, LANGUAGE })}
+						{Engine.component(ForgetPasswordBox, { BOXES, ERRORS, LANGUAGE })}
+						{Engine.component(LoginBox, { BOXES, ERRORS, LANGUAGE })}
+						{Engine.component(Register, { BOXES, ERRORS, LANGUAGE })}
+						{Engine.component(SearchBox, { BOXES, ERRORS, LANGUAGE })}
 					</div>
 				</nav>
 			</nav>
