@@ -1,10 +1,7 @@
 import React, { Component } from 'react';
 import { connect } from 'react-redux';
 import { Button } from 'react-bootstrap';
-import MediaQuery from 'react-responsive';
 import browser from '../image/gui/search-icon.svg';
-import logo from '../logo.svg';
-import node from '../node.png';
 import { Engine } from '../common/engine';
 import { Helper } from '../common/js-funcs';
 import { getLabel } from '../common/label';
@@ -20,24 +17,16 @@ class Footer extends Component {
 		return (
 			<footer>
 				<div className="footer">
-					<MediaQuery maxWidth={768}>
-						<div className="react-icon" title={getLabel('host_info')}>
-							<img className="rotate-on" alt="" src={logo} height="20" />
-						</div>
-						<div className="react-icon" title="NodeJs">
-							<img alt="" src={node} height="20" />
-						</div>
-					</MediaQuery>
-					{this.props.USERSIGNEDIN ? (<Button bsSize="small" className="custom-button-white" onClick={() => {UserFunction.signOut()}}>{getLabel('c_sign_out')}</Button>) : (<Button bsSize="small" className="custom-button-white" onClick={() => Helper.launchMenu('login')}>{getLabel('c_sign_in')}</Button>)}
-					<a href="" onClick={e => {e.preventDefault(); Helper.launchMenu('search')}}><img alt='' title={getLabel('c_search')} className={`svg-icon ${off}`} src={browser} /></a>
+					{this.props.USERSIGNEDIN ? (<Button bsSize="small" className="custom-button-white" onClick={() => { UserFunction.signOut() }}>{getLabel('c_sign_out')}</Button>) : (<Button bsSize="small" className="custom-button-white" onClick={() => Helper.launchMenu('login')}>{getLabel('c_sign_in')}</Button>)}
+					<a href="" onClick={e => { e.preventDefault(); Helper.launchMenu('search') }}><img alt='' title={getLabel('c_search')} className={`svg-icon ${off}`} src={browser} /></a>
 					<div className="fb-like" data-href="http://esaleshome.com" data-layout="button" data-action="like" data-show-faces="true" data-share="true"></div>
 				</div>
 				<section className={siteMsg}>
-					{Engine.component(MessageBox, {messageCenter: this.props.messageCenter})}
+					{Engine.component(MessageBox, { messageCenter: this.props.messageCenter })}
 				</section>
 				{Engine.component(Menus)}
 				{Engine.component(LargeImages)}
-			</footer>	
+			</footer>
 		);
 	};
 };
