@@ -1,4 +1,3 @@
-import { isIE } from 'react-device-detect';
 import React from 'react';
 import ReactDOM from 'react-dom';
 import { createStore, applyMiddleware } from 'redux';
@@ -12,27 +11,21 @@ import { Engine } from './common/engine';
 import registerServiceWorker from './registerServiceWorker';
 
 const sagaMiddleware = createSagaMiddleware(),
-store = createStore(
-	RootReducer,
-	applyMiddleware(sagaMiddleware)
-);
+	store = createStore(
+		RootReducer,
+		applyMiddleware(sagaMiddleware)
+	);
 
 Engine.register(store);
 
 sagaMiddleware.run(RootSaga);
 
-ReactDOM.render(
-	isIE ? (
-		<section className="not-signed-in">
-			<span>Internet Explorer is not supported!</span>
-		</section>
-	) : (
-		<Provider store={store}>
-			<CookiesProvider>
-				<Advertise />
-			</CookiesProvider>
-		</Provider>
-	),
-	document.getElementById('esaleshome-root')
+ReactDOM.render((
+	<Provider store={store}>
+		<CookiesProvider>
+			<Advertise />
+		</CookiesProvider>
+	</Provider>
+), document.getElementById('esaleshome-root')
 );
 registerServiceWorker();
